+++ src/lib/prompts/followup/follow-up-patching-with-tools.js
@@ -2,35 +2,11 @@ import { ICON_CONFIGURATION } from '../new-extension/one-shot/shared-content.js'
 import {
   CHROME_API_SEARCH_TOOL,
   WEB_SCRAPING_TOOL,
-  FILE_DELETE_TOOL
+  FILE_DELETE_TOOL,
+  buildToolDescriptions
 } from './tools/index.js';
 
-export { CHROME_API_SEARCH_TOOL, WEB_SCRAPING_TOOL, FILE_DELETE_TOOL };
-
-/**
- * Builds conditional tool descriptions based on enabled tools
- * @param {string[]} enabledTools - Array of tool names to enable
- * @returns {string} - XML-formatted tool descriptions or empty string
- */
-export function buildToolDescriptions(enabledTools = []) {
-  if (!enabledTools || enabledTools.length === 0) {
-    return '';
-  }
-
-  const toolMap = {
-    'chrome_api_search': CHROME_API_SEARCH_TOOL,
-    'web_scraping': WEB_SCRAPING_TOOL,
-    'delete_file': FILE_DELETE_TOOL
-  };
-
-  const sections = enabledTools
-    .filter(tool => toolMap[tool])
-    .map(tool => toolMap[tool]);
-
-  if (sections.length === 0) return '';
-
-  return `<available_tools>\n${sections.join('\n')}\n</available_tools>`;
-}
+export { CHROME_API_SEARCH_TOOL, WEB_SCRAPING_TOOL, FILE_DELETE_TOOL, buildToolDescriptions };
 
 export const FOLLOW_UP_PATCH_PROMPT_WITH_TOOLS = `
 <s>
